+++ changelog.md
@@ -1,5 +1,8 @@
-# 2.6.3
+# 2.7.0
 
+- Added **Icon Effects**
+- Added Reset percentage on save
+- Added contribute button to languages tab
 - Fixed Input nodes not working on macOS and android (with floating keyboards)
 - Fixed colour gradients not saving
 - Fixed keybind to open menu not working sometimes